Fixing build errors and nextjs version
